Patch 0969
・wis送信パケットをパケット仕様に合わせて変更 (src/char/inter.c src/map/clif.* src/map/inter.c doc/inter_server_packet.txt doc/client_packet.txt)
・少しリファクタ (src/common/mmo.h src/map/clif.c)
・新仕様のパーティーブッキングはパーティーリーダーでないと登録が出来ないよう修正 (src/map/booking.c)
・パケットの解析結果を記載と修正 (doc/client_packet.txt)
・キャラ作成時のパラメータの型を一部変更 (src/char/char.c src/char/txt/chardb_txt.* src/char/sql/chardb_sql.*)
・2012-06-18aRagexeREまでのパケット長の追記と一部クリーンアップと修正 (db/packet_db.txt)

+++ src/char/char.c
@@ -2569,8 +2569,8 @@ int parse_char(int fd)
 				short dex   = RFIFOB(fd,30);
 				short luk   = RFIFOB(fd,31);
 				unsigned char slot = RFIFOB(fd,32);
-				int hair_color = RFIFOW(fd,33);
-				int hair = RFIFOW(fd,35);
+				short hair_color = RFIFOW(fd,33);
+				short hair = RFIFOW(fd,35);
 #else
 				unsigned char *name  = RFIFOP(fd,2);
 				short str   = 5;
@@ -2580,8 +2580,8 @@ int parse_char(int fd)
 				short dex   = 5;
 				short luk   = 5;
 				unsigned char slot = RFIFOB(fd,26);
-				int hair_color = RFIFOW(fd,27);
-				int hair = RFIFOW(fd,29);
+				short hair_color = RFIFOW(fd,27);
+				short hair = RFIFOW(fd,29);
 #endif
 				int flag=0x04;
 				int i = 0;
